Switch to 3-clause BSD license.
This makes minipb use the same license as official protobuf.

LGPL has an implication that the code must be swappable by the user, which might be undesireable for some devs who don't wish to ship software in incomplete/user-modifiable form. Besides that, it looks like that the chance will be pretty slim that proprietary software developers will start to contribute to this project anyway.